Updated ARCore SDK and PlayServices as well as ARFoundation to Candidates instead of bintray, removed ARMock (#205)

+++ FunctionalTests/AR/ARCore/Assets/GoogleARCore/SDK/Scripts/Api/Wrappers/SessionApi.cs
@@ -1,7 +1,7 @@
 //-----------------------------------------------------------------------
 // <copyright file="SessionApi.cs" company="Google">
 //
-// Copyright 2017 Google Inc. All Rights Reserved.
+// Copyright 2017 Google LLC. All Rights Reserved.
 //
 // Licensed under the Apache License, Version 2.0 (the "License");
 // you may not use this file except in compliance with the License.
@@ -45,36 +45,40 @@ namespace GoogleARCoreInternal
 
         public void ReportEngineType()
         {
-            ExternApi.ArSession_reportEngineType(m_NativeSession.SessionHandle, "Unity", Application.unityVersion);
+            ExternApi.ArSession_reportEngineType(
+                m_NativeSession.SessionHandle, "Unity", Application.unityVersion);
         }
 
-        public bool SetConfiguration(ARCoreSessionConfig sessionConfig)
+        public void GetSupportedCameraConfigurationsWithFilter(
+            ARCoreCameraConfigFilter cameraConfigFilter,
+            IntPtr cameraConfigListHandle, List<IntPtr> supportedCameraConfigHandles,
+            List<CameraConfig> supportedCameraConfigs, DeviceCameraDirection cameraFacingDirection)
         {
-            IntPtr configHandle = m_NativeSession.SessionConfigApi.Create();
-            m_NativeSession.SessionConfigApi.UpdateApiConfigWithArCoreSessionConfig(configHandle, sessionConfig);
-
-            bool ret = ExternApi.ArSession_configure(m_NativeSession.SessionHandle, configHandle) == 0;
-            m_NativeSession.SessionConfigApi.Destroy(configHandle);
-
-            return ret;
-        }
-
-        public void GetSupportedCameraConfigurations(IntPtr cameraConfigListHandle,
-            List<IntPtr> supportedCameraConfigHandles, List<CameraConfig> supportedCameraConfigs)
-        {
-            ExternApi.ArSession_getSupportedCameraConfigs(
-                m_NativeSession.SessionHandle, cameraConfigListHandle);
+            IntPtr cameraConfigFilterHandle =
+                m_NativeSession.CameraConfigFilterApi.Create(cameraConfigFilter);
+            ExternApi.ArSession_getSupportedCameraConfigsWithFilter(m_NativeSession.SessionHandle,
+                cameraConfigFilterHandle, cameraConfigListHandle);
+            m_NativeSession.CameraConfigFilterApi.Destroy(cameraConfigFilterHandle);
 
             supportedCameraConfigHandles.Clear();
             supportedCameraConfigs.Clear();
             int listSize = m_NativeSession.CameraConfigListApi.GetSize(cameraConfigListHandle);
-            Debug.LogFormat("Found {0} camera configs.", listSize);
 
             for (int i = 0; i < listSize; i++)
             {
                 IntPtr cameraConfigHandle = m_NativeSession.CameraConfigApi.Create();
-                m_NativeSession.CameraConfigListApi.GetItemAt(cameraConfigListHandle, i, cameraConfigHandle);
-                Debug.LogFormat("Config {0} with handle {1}", i, cameraConfigHandle);
+                m_NativeSession.CameraConfigListApi.GetItemAt(
+                    cameraConfigListHandle, i, cameraConfigHandle);
+
+                // Skip camera config that has a different camera facing direction.
+                DeviceCameraDirection configDirection =
+                    m_NativeSession.CameraConfigApi.GetFacingDirection(cameraConfigHandle)
+                        .ToDeviceCameraDirection();
+                if (configDirection != cameraFacingDirection)
+                {
+                    continue;
+                }
+
                 supportedCameraConfigHandles.Add(cameraConfigHandle);
                 supportedCameraConfigs.Add(_CreateCameraConfig(cameraConfigHandle));
             }
@@ -82,12 +86,20 @@ namespace GoogleARCoreInternal
 
         public ApiArStatus SetCameraConfig(IntPtr cameraConfigHandle)
         {
-            return ExternApi.ArSession_setCameraConfig(m_NativeSession.SessionHandle, cameraConfigHandle);
+            return ExternApi.ArSession_setCameraConfig(
+                m_NativeSession.SessionHandle, cameraConfigHandle);
         }
 
         public CameraConfig GetCameraConfig()
         {
             IntPtr cameraConfigHandle = m_NativeSession.CameraConfigApi.Create();
+
+            if (InstantPreviewManager.IsProvidingPlatform)
+            {
+                InstantPreviewManager.LogLimitedSupportMessage("access camera config");
+                return new CameraConfig();
+            }
+
             ExternApi.ArSession_getCameraConfig(m_NativeSession.SessionHandle, cameraConfigHandle);
             CameraConfig currentCameraConfig = _CreateCameraConfig(cameraConfigHandle);
             m_NativeSession.CameraConfigApi.Destroy(cameraConfigHandle);
@@ -97,13 +109,15 @@ namespace GoogleARCoreInternal
         public void GetAllTrackables(List<Trackable> trackables)
         {
             IntPtr listHandle = m_NativeSession.TrackableListApi.Create();
-            ExternApi.ArSession_getAllTrackables(m_NativeSession.SessionHandle, ApiTrackableType.BaseTrackable, listHandle);
+            ExternApi.ArSession_getAllTrackables(
+                m_NativeSession.SessionHandle, ApiTrackableType.BaseTrackable, listHandle);
 
             trackables.Clear();
             int count = m_NativeSession.TrackableListApi.GetCount(listHandle);
             for (int i = 0; i < count; i++)
             {
-                IntPtr trackableHandle = m_NativeSession.TrackableListApi.AcquireItem(listHandle, i);
+                IntPtr trackableHandle =
+                    m_NativeSession.TrackableListApi.AcquireItem(listHandle, i);
 
                 Trackable trackable = m_NativeSession.TrackableFactory(trackableHandle);
                 if (trackable != null)
@@ -143,32 +157,36 @@ namespace GoogleARCoreInternal
                     break;
             }
 
-            ExternApi.ArSession_setDisplayGeometry(m_NativeSession.SessionHandle, androidOrientation, width, height);
+            ExternApi.ArSession_setDisplayGeometry(
+                m_NativeSession.SessionHandle, androidOrientation, width, height);
         }
 
         public Anchor CreateAnchor(Pose pose)
         {
             IntPtr poseHandle = m_NativeSession.PoseApi.Create(pose);
             IntPtr anchorHandle = IntPtr.Zero;
-            ExternApi.ArSession_acquireNewAnchor(m_NativeSession.SessionHandle, poseHandle, ref anchorHandle);
+            ExternApi.ArSession_acquireNewAnchor(
+                m_NativeSession.SessionHandle, poseHandle, ref anchorHandle);
             var anchorResult = Anchor.Factory(m_NativeSession, anchorHandle);
             m_NativeSession.PoseApi.Destroy(poseHandle);
             return anchorResult;
         }
 
-        public ApiArStatus CreateCloudAnchor(IntPtr platformAnchorHandle, out IntPtr cloudAnchorHandle)
+        public ApiArStatus CreateCloudAnchor(
+            IntPtr platformAnchorHandle, out IntPtr cloudAnchorHandle)
         {
             cloudAnchorHandle = IntPtr.Zero;
-            var result = ExternApi.ArSession_hostAndAcquireNewCloudAnchor(m_NativeSession.SessionHandle,
-                platformAnchorHandle, ref cloudAnchorHandle);
+            var result =
+                ExternApi.ArSession_hostAndAcquireNewCloudAnchor(
+                    m_NativeSession.SessionHandle, platformAnchorHandle, ref cloudAnchorHandle);
             return result;
         }
 
         public ApiArStatus ResolveCloudAnchor(String cloudAnchorId, out IntPtr cloudAnchorHandle)
         {
             cloudAnchorHandle = IntPtr.Zero;
-            return ExternApi.ArSession_resolveAndAcquireNewCloudAnchor(m_NativeSession.SessionHandle,
-                cloudAnchorId, ref cloudAnchorHandle);
+            return ExternApi.ArSession_resolveAndAcquireNewCloudAnchor(
+                m_NativeSession.SessionHandle, cloudAnchorId, ref cloudAnchorHandle);
         }
 
         private CameraConfig _CreateCameraConfig(IntPtr cameraConfigHandle)
@@ -177,13 +195,19 @@ namespace GoogleARCoreInternal
             int imageHeight = 0;
             int textureWidth = 0;
             int textureHeight = 0;
-            m_NativeSession.CameraConfigApi.GetImageDimensions(cameraConfigHandle,
-                out imageWidth, out imageHeight);
-            m_NativeSession.CameraConfigApi.GetTextureDimensions(cameraConfigHandle,
-                out textureWidth, out textureHeight);
+            int minFps = 0;
+            int maxFps = 0;
+            CameraConfigDepthSensorUsages depthSensorUsage =
+                m_NativeSession.CameraConfigApi.GetDepthSensorUsage(cameraConfigHandle);
+            m_NativeSession.CameraConfigApi.GetImageDimensions(
+                cameraConfigHandle, out imageWidth, out imageHeight);
+            m_NativeSession.CameraConfigApi.GetTextureDimensions(
+                cameraConfigHandle, out textureWidth, out textureHeight);
+            m_NativeSession.CameraConfigApi.GetFpsRange(
+                cameraConfigHandle, out minFps, out maxFps);
 
             return new CameraConfig(new Vector2(imageWidth, imageHeight),
-                new Vector2(textureWidth, textureHeight));
+                new Vector2(textureWidth, textureHeight), minFps, maxFps, depthSensorUsage);
         }
 
         private struct ExternApi
@@ -193,38 +217,46 @@ namespace GoogleARCoreInternal
             public static extern int ArSession_configure(IntPtr sessionHandle, IntPtr config);
 
             [AndroidImport(ApiConstants.ARCoreNativeApi)]
-            public static extern void ArSession_getSupportedCameraConfigs(IntPtr sessionHandle, 
+            public static extern void ArSession_getSupportedCameraConfigsWithFilter(
+                IntPtr sessionHandle, IntPtr cameraConfigFilterHandle,
                 IntPtr cameraConfigListHandle);
 
             [AndroidImport(ApiConstants.ARCoreNativeApi)]
-            public static extern ApiArStatus ArSession_setCameraConfig(IntPtr sessionHandle, IntPtr cameraConfigHandle);
+            public static extern ApiArStatus ArSession_setCameraConfig(
+                IntPtr sessionHandle, IntPtr cameraConfigHandle);
 
             [AndroidImport(ApiConstants.ARCoreNativeApi)]
-            public static extern void ArSession_getCameraConfig(IntPtr sessionHandle, IntPtr cameraConfigHandle);
+            public static extern void ArSession_getCameraConfig(
+                IntPtr sessionHandle, IntPtr cameraConfigHandle);
 
             [AndroidImport(ApiConstants.ARCoreNativeApi)]
-            public static extern void ArSession_getAllTrackables(IntPtr sessionHandle, ApiTrackableType filterType,
-                IntPtr trackableList);
+            public static extern void ArSession_getAllTrackables(
+                IntPtr sessionHandle, ApiTrackableType filterType, IntPtr trackableList);
 
             [AndroidImport(ApiConstants.ARCoreNativeApi)]
-            public static extern void ArSession_setDisplayGeometry(IntPtr sessionHandle, int rotation, int width,
-                int height);
+            public static extern void ArSession_setDisplayGeometry(
+                IntPtr sessionHandle, int rotation, int width, int height);
 
             [AndroidImport(ApiConstants.ARCoreNativeApi)]
-            public static extern int ArSession_acquireNewAnchor(IntPtr sessionHandle, IntPtr poseHandle,
-                ref IntPtr anchorHandle);
+            public static extern int ArSession_acquireNewAnchor(
+                IntPtr sessionHandle, IntPtr poseHandle, ref IntPtr anchorHandle);
 #pragma warning restore 626
-            [DllImport(ApiConstants.ARCoreNativeApi)]
-            public static extern void ArSession_reportEngineType(IntPtr sessionHandle, string engineType,
-                string engineVersion);
 
             [DllImport(ApiConstants.ARCoreNativeApi)]
-            public static extern ApiArStatus ArSession_hostAndAcquireNewCloudAnchor(IntPtr sessionHandle,
-                IntPtr anchorHandle, ref IntPtr cloudAnchorHandle);
+            public static extern void ArSession_reportEngineType(
+                IntPtr sessionHandle, string engineType, string engineVersion);
+
+            [DllImport(ApiConstants.ARCoreNativeApi)]
+            public static extern ApiArStatus ArSession_hostAndAcquireNewCloudAnchor(
+                IntPtr sessionHandle,
+                IntPtr anchorHandle,
+                ref IntPtr cloudAnchorHandle);
 
             [DllImport(ApiConstants.ARCoreNativeApi)]
             public static extern ApiArStatus ArSession_resolveAndAcquireNewCloudAnchor(
-                IntPtr sessionHandle, String cloudAnchorId,  ref IntPtr cloudAnchorHandle);
+                IntPtr sessionHandle,
+                String cloudAnchorId,
+                ref IntPtr cloudAnchorHandle);
         }
     }
 }
